script/photos-setup/requirements.txt
opencv_python==4.10.0.84
Pillow==10.4.0
